Moved documentation and enhanced it
