+++ Docs/01 Basic tutorials/04 Elements overview.md
@@ -1,0 +1,85 @@
+
+# WIP
+
+Every listed element exists and is perfectly useable, if not marked as WIP.\
+The documentation itself is WIP.
+
+
+# Elements overview
+
+This tutorial gives you an overview over all the available SwiftGUI-elements.
+
+Most of these elements can do much more than I show in this tutorial, so it pays off to read the corresponding element-tutorials.
+
+# Widget-elements
+This chapter covers all "non-combined" elements, which contain only a single tkinter-widget.
+
+## Text
+Aliases: T, Label
+
+Displays a text. Nothing more, nothing less.
+
+**Elements that have any text in them usually contain all options of this element**, so they are explained a bit more detailed.
+
+## Button
+
+
+## Checkbox
+Aliases: Check, Checkbutton
+
+
+## Frame
+Aliases: Column
+
+
+## Input
+Aliases: In, Entry
+
+
+## Separators (HorizontalSeparator, VerticalSeparator)
+Aliases: HSep, VSep
+
+
+## Spacer
+Aliases: S
+
+To be exact, this is an extended element, but it's so basic, let's keep it here.
+
+## Listbox
+
+
+## TKContainer
+Aliases: TKWidget
+
+
+## TextField
+Aliases: Multiline
+
+
+## Treeview (WIP)
+
+
+## Table
+
+
+## Notebook
+Aliases: TabGroup
+
+
+# Elements with extended functionality
+SwiftGUI offers a couple of elements that contain a single tkinter-widget, but change its functionality.
+
+## FileBrowseButton
+A button that opens a file-dialogue when clicked.
+
+
+## ColorChooserButton
+A button that opens a color-chooser-menu when clicked.
+
+
+
+# Combined elements
+SwiftGUI offeres a couple of "combined" elements that consist of multiple SwiftGUI-elements.
+
+## Form (WIP)
+
